requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
pytest
httpx
faiss-cpu
